+++ app/src/main/java/com/surveasy/surveasy/presentation/main/survey/SurveyActivity.kt
@@ -1,5 +1,6 @@
 package com.surveasy.surveasy.presentation.main.survey
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.navigation.NavController
@@ -7,6 +8,7 @@ import androidx.navigation.fragment.NavHostFragment
 import com.surveasy.surveasy.R
 import com.surveasy.surveasy.databinding.ActivitySurveyBinding
 import com.surveasy.surveasy.presentation.base.BaseActivity
+import com.surveasy.surveasy.presentation.main.MainActivity
 import com.surveasy.surveasy.presentation.util.IntentId.SURVEY_ID
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -16,7 +18,16 @@ class SurveyActivity : BaseActivity<ActivitySurveyBinding>(ActivitySurveyBinding
     private lateinit var navHostFragment: NavHostFragment
     private lateinit var navController: NavController
 
-    override fun initEventObserver() = Unit
+    override fun initEventObserver() {
+        repeatOnStarted {
+            viewModel.events.collect {
+                when(it){
+                    is SurveyEvents.NavigateToList -> toList()
+                    else -> Unit
+                }
+            }
+        }
+    }
 
     override fun initView() {
         navHostFragment =
@@ -28,4 +39,7 @@ class SurveyActivity : BaseActivity<ActivitySurveyBinding>(ActivitySurveyBinding
         val sId = intent.getIntExtra(SURVEY_ID, -1)
         viewModel.setSurveyId(sId)
     }
+
+    private fun toList() = finish()
+
 }
