+++ app/src/main/res/layout/fragment_survey_done.xml
@@ -42,35 +42,20 @@
             app:layout_constraintVertical_chainStyle="packed" />
 
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_mypage"
-            style="@style/TextMediumBold"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@drawable/white_button"
-            android:onClick="@{() -> vm.navigateToMy()}"
-            android:paddingHorizontal="12dp"
-            android:text="마이페이지로 가기"
-            android:textColor="@color/main_green"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/btn_list"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_done"
-            app:layout_constraintVertical_chainStyle="packed" />
-
-        <androidx.appcompat.widget.AppCompatButton
             android:id="@+id/btn_list"
             style="@style/TextMediumBold"
-            android:layout_width="wrap_content"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:background="@drawable/register_button"
             android:onClick="@{() -> vm.navigateToList()}"
             android:paddingHorizontal="29dp"
-            android:text="목록으로 가기"
+            app:getReward="@{vm.uiState.reward}"
             android:textColor="@color/white"
-            app:layout_constraintBottom_toBottomOf="@id/btn_mypage"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/btn_mypage"
-            app:layout_constraintTop_toTopOf="@id/btn_mypage"
+            app:layout_constraintTop_toBottomOf="@id/tv_done"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginTop="60dp"
+            app:layout_constraintStart_toStartOf="@id/tv_done"
+            app:layout_constraintEnd_toEndOf="@id/tv_done"
             app:layout_constraintVertical_chainStyle="packed" />
